requirements.txt
Django==3.0.5
djangorestframework==3.11.0
huey==2.2.0
psycopg2==2.8.5
ruuvitag-sensor==1.0.1
bleson==0.0.10